chore(plugin): extract same utils to shared
