+++ .changeset/tender-buttons-tie.md
@@ -1,0 +1,5 @@
+---
+'@modern-js/plugin': patch
+---
+
+chore(plugin): extract same utils to shared
